Sandboxes/index-factory/worker/requirements.txt
celery[redis]==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
qdrant-client==1.12.1
open-clip-torch==2.29.0
sentence-transformers==3.3.1
Pillow==11.0.0
tiktoken==0.8.0
structlog==24.4.0
torch==2.5.1
torchvision==0.20.1
